Better Error Messages on not Windows/Linux
MacOS builds of bass-sys right now attempt to use the Linux bindings
and then fail with obscure errors about file formats. This change adds
better checking for that and checks the compile target of the crate,
not the build script. This comes up when cross compiling.

See: https://kazlauskas.me/entries/writing-proper-buildrs-scripts.html